+++ src/io/infer_format/dispatch.rs
@@ -1,5 +1,5 @@
 #![allow(clippy::type_complexity, clippy::large_enum_variant)]
-use std::{fmt::Debug, fs, io, marker::PhantomData, path::{self, Path}};
+use std::{fmt::Debug, fs, io::{self, Cursor}, marker::PhantomData, path::{self, Path}};
 
 use flate2::read::GzDecoder;
 use mzpeaks::{prelude::FeatureLike, CentroidLike, CentroidPeak, DeconvolutedCentroidLike, DeconvolutedPeak, KnownCharge};
@@ -12,7 +12,7 @@ use crate::{io::{Generic3DIonMobilityFrameSource,
     PreBufferedStream,
     RandomAccessIonMobilityFrameIterator},
     spectrum::MultiLayerIonMobilityFrame};
-use crate::io::{traits::{MZFileReader, RandomAccessSpectrumIterator, SpectrumSource}, RestartableGzDecoder};
+use crate::{io::{traits::{MZFileReader, RandomAccessSpectrumIterator, SpectrumSource}, MemorySpectrumSource, RestartableGzDecoder}, meta::FileMetadataConfig, prelude::SeekRead};
 #[cfg(feature = "mzmlb")]
 pub use crate::io::mzmlb::MzMLbReaderType;
 
@@ -152,6 +152,11 @@ impl<C: CentroidLike + From<CentroidPeak> + BuildFromArrayMap, D: DeconvolutedCe
         Ok(reader)
     }
 
+    /// See [`MZReaderType::open_read_seek_generic`]
+    pub fn from_read_seek_generic(self, source: Box<dyn SeekRead>) -> Result<MZReaderType<Box<dyn SeekRead>, C, D>, io::Error> {
+        MZReaderType::<Box<dyn SeekRead>, C, D>::open_read_seek_generic(source)
+    }
+
     /// Create a reader from a type that supports [`io::Read`].
     ///
     /// This will internally wrap the file in a [`PreBufferedStream`] for metadata
@@ -170,6 +175,26 @@ impl<C: CentroidLike + From<CentroidPeak> + BuildFromArrayMap, D: DeconvolutedCe
         }?;
         reader.get_mut().set_detail_level(self.detail_level);
         Ok(reader)
+    }
+
+    /// See [`MZReaderType::open_read_generic`]
+    pub fn from_read_generic(self, source: Box<dyn io::Read>) -> io::Result<StreamingSpectrumIterator<C, D, MultiLayerSpectrum<C, D>, MZReaderType<PreBufferedStream<Box<dyn io::Read>>, C, D>>> {
+        MZReaderType::<PreBufferedStream<fs::File>, C, D>::open_read_generic(source)
+    }
+
+    /// Create a "reader" from an existing collection of spectra in memory.
+    ///
+    /// Extra metadata may be provided if it is available, but it is not required.
+    ///
+    /// # Note
+    /// This uses [`MemorySpectrumSource`] internally, which requires the spectrum data structure
+    /// implement [`Clone`]. All of the [`SpectrumLike`] types here derive `Clone`, but the peak types
+    /// must also implement [`Clone`] too.
+    ///
+    /// All spectra are kept in memory, so keep this in mind if you are passing a large collection!
+    pub fn from_spectra(self, source: impl IntoIterator<Item = MultiLayerSpectrum<C, D>>, metadata: Option<FileMetadataConfig>) -> MZReaderType<Cursor<&'static [u8]>, C, D>
+            where C: Clone + Send + Sync + 'static, D: Clone + Send + Sync + 'static {
+        MZReaderType::<Cursor<&'static [u8]>, C, D>::open_spectra(source, metadata)
     }
 }
 
@@ -259,6 +284,12 @@ impl<R: io::Read + io::Seek,
         if gzipped {
             return Err(io::Error::new(io::ErrorKind::Unsupported, "This method does not support gzipped streams"))
         }
+        MZReaderType::create_from_read_seek(stream, fmt)
+    }
+
+    /// Shared plumbing for all generic read-seek openers, so that `n+1`th format changes
+    /// are repeated as few times as possible
+    fn create_from_read_seek(stream: R, fmt: MassSpectrometryFormat) -> io::Result<Self> {
         match fmt {
             #[cfg(feature = "mgf")]
             MassSpectrometryFormat::MGF => Ok(Self::MGF(MGFReaderType::new_indexed(stream))),
@@ -271,6 +302,27 @@ impl<R: io::Read + io::Seek,
     }
 
     /// Create a reader from a type that supports [`io::Read`] and
+    /// [`io::Seek`], generic over whether or not the stream is gzip-compressed
+    /// or not, at the cost of an extra layer of indirection.
+    ///
+    /// # Note
+    /// Not all formats can be read from an `io` type, these will
+    /// fail to open and an error will be returned
+    ///
+    /// # Performance
+    /// See [`MZReaderType::open_gzipped_read_seek`] for more details
+    pub fn open_read_seek_generic(mut stream: Box<dyn SeekRead>) -> io::Result<MZReaderType<Box<dyn SeekRead>, C, D>> {
+        let (fmt, gzipped) = infer_from_stream(&mut stream)?;
+        if gzipped {
+            log::trace!("Reading gzipped generic seekable stream");
+            let stream = Box::new(RestartableGzDecoder::new(io::BufReader::new(stream))) as Box<dyn SeekRead>;
+            MZReaderType::create_from_read_seek(stream, fmt)
+        } else {
+            MZReaderType::open_read_seek(stream)
+        }
+    }
+
+    /// Create a reader from a type that supports [`io::Read`] and
     /// [`io::Seek`] that is gzip-compressed.
     ///
     /// # Note
@@ -278,21 +330,35 @@ impl<R: io::Read + io::Seek,
     ///
     /// Not all formats can be read from an `io` type, these will
     /// fail to open and an error will be returned
+    ///
+    /// # Performance
+    /// A gzip compressed stream is very slow to do random access on, especially seeking backwards.
+    /// This method is provided as a convenience, but special care should be taken if you anticipate doing things
+    /// like searching over time or backtracking to find precursors. Often, a different algorithm using
+    /// the non-`io::Seek`-requiring method helps in these scenarios.
     pub fn open_gzipped_read_seek(mut stream: R) -> io::Result<MZReaderType<RestartableGzDecoder<io::BufReader<R>>, C, D>> {
         let (fmt, gzipped) = infer_from_stream(&mut stream)?;
         if !gzipped {
             return Err(io::Error::new(io::ErrorKind::Unsupported, "This method does not support non-gzipped streams"))
         }
         let stream = RestartableGzDecoder::new(io::BufReader::new(stream));
-        match fmt {
-            #[cfg(feature = "mgf")]
-            MassSpectrometryFormat::MGF => Ok(MZReaderType::MGF(MGFReaderType::new_indexed(stream))),
-            #[cfg(feature = "mzml")]
-            MassSpectrometryFormat::MzML => Ok(MZReaderType::MzML(MzMLReaderType::new_indexed(stream))),
-            _ => {
-                Err(io::Error::new(io::ErrorKind::Unsupported, format!("This method does not support {fmt}")))
-            }
-        }
+        MZReaderType::create_from_read_seek(stream, fmt)
+    }
+
+    /// Create a "reader" from an existing collection of spectra in memory.
+    ///
+    /// Extra metadata may be provided if it is available, but it is not required.
+    ///
+    /// # Note
+    /// This uses [`MemorySpectrumSource`] internally, which requires the spectrum data structure
+    /// implement [`Clone`]. All of the [`SpectrumLike`] types here derive `Clone`, but the peak types
+    /// must also implement [`Clone`] too.
+    ///
+    /// All spectra are kept in memory, so keep this in mind if you are passing a large collection!
+    pub fn open_spectra(source: impl IntoIterator<Item = MultiLayerSpectrum<C, D>>, metadata: Option<FileMetadataConfig>) -> MZReaderType<Cursor<&'static [u8]>, C, D>
+            where C: Clone + Send + Sync + 'static, D: Clone + Send + Sync + 'static {
+        let source = MemorySpectrumSource::new_with_metadata(source.into_iter().collect(), metadata.unwrap_or_default());
+        MZReaderType::Unknown(Box::new(source), PhantomData)
     }
 }
 
@@ -339,6 +405,21 @@ impl<R: io::Read,
      C: CentroidLike + From<CentroidPeak> + BuildFromArrayMap,
      D: DeconvolutedCentroidLike + From<DeconvolutedPeak> + BuildFromArrayMap> MZReaderType<PreBufferedStream<R>, C, D> {
 
+    /// Shared plumbing for all generic read openers, so that `n+1`th format changes
+    /// are repeated as few times as possible
+    fn create_from_prebuffered_reader(stream: PreBufferedStream<R>, fmt: MassSpectrometryFormat) -> io::Result<StreamingSpectrumIterator<C, D, MultiLayerSpectrum<C, D>, MZReaderType<PreBufferedStream<R>, C, D>>> {
+        let reader = match fmt {
+            #[cfg(feature = "mgf")]
+            MassSpectrometryFormat::MGF => MZReaderType::MGF(MGFReaderType::new(stream)),
+            #[cfg(feature = "mzml")]
+            MassSpectrometryFormat::MzML => MZReaderType::MzML(MzMLReaderType::new(stream)),
+            _ => {
+                return Err(io::Error::new(io::ErrorKind::Unsupported, format!("This method does not support {fmt}")))
+            }
+        };
+        Ok(StreamingSpectrumIterator::new(reader))
+    }
+
     /// Create a reader from a type that supports [`io::Read`].
     ///
     /// This will internally wrap the file in a [`PreBufferedStream`] for metadata
@@ -357,16 +438,7 @@ impl<R: io::Read,
             return Err(io::Error::new(io::ErrorKind::Unsupported, "This method does not support gzipped streams"))
         }
 
-        let reader = match fmt {
-            #[cfg(feature = "mgf")]
-            MassSpectrometryFormat::MGF => Self::MGF(MGFReaderType::new(stream)),
-            #[cfg(feature = "mzml")]
-            MassSpectrometryFormat::MzML => Self::MzML(MzMLReaderType::new(stream)),
-            _ => {
-                return Err(io::Error::new(io::ErrorKind::Unsupported, format!("This method does not support {fmt}")))
-            }
-        };
-        Ok(StreamingSpectrumIterator::new(reader))
+        MZReaderType::create_from_prebuffered_reader(stream, fmt)
     }
 
     /// Create a reader from a type that supports [`io::Read`] is gzip-compressed.
@@ -388,25 +460,30 @@ impl<R: io::Read,
         if !gzipped {
             return Err(io::Error::new(io::ErrorKind::Unsupported, "This method only supports gzipped streams"))
         }
-        let mut stream = PreBufferedStream::new(GzDecoder::new(stream))?;
+        let stream = PreBufferedStream::new(GzDecoder::new(stream))?;
+        MZReaderType::create_from_prebuffered_reader(stream, fmt)
+    }
 
-        let reader = match fmt {
-            #[cfg(feature = "mgf")]
-            MassSpectrometryFormat::MGF => MZReaderType::MGF(MGFReaderType::new(stream)),
-            #[cfg(feature = "mzml")]
-            MassSpectrometryFormat::MzML => MZReaderType::MzML(MzMLReaderType::new(stream)),
-            _ => {
-                return Err(io::Error::new(io::ErrorKind::Unsupported, format!("This method does not support {fmt}")))
-            }
-        };
-        Ok(StreamingSpectrumIterator::new(reader))
+    /// Create a reader from a type that supports [`io::Read`] generic over gzipped
+    /// compression, at the cost of another layer of indirection.
+    ///
+    /// See [`MZReaderType::open_read`] for more details.
+    pub fn open_read_generic(stream: Box<dyn io::Read>) -> io::Result<StreamingSpectrumIterator<C, D, MultiLayerSpectrum<C, D>, MZReaderType<PreBufferedStream<Box<dyn io::Read>>, C, D>>> {
+        let mut stream = PreBufferedStream::new(stream)?;
+        let (fmt, gzipped) = infer_from_stream(&mut stream)?;
+        if gzipped {
+            let decoder = Box::new(GzDecoder::new(stream)) as Box<dyn io::Read>;
+            let mut stream = PreBufferedStream::new(decoder)?;
+            MZReaderType::create_from_prebuffered_reader(stream, fmt)
+        } else {
+            MZReaderType::create_from_prebuffered_reader(stream, fmt)
+        }
     }
 
     /// See [`MZReaderType::open_read`].
     ///
     /// This function lets the caller specify the prebuffering size for files with large
     /// headers that exceed the default buffer size.
-    #[allow(unreachable_code)]
     pub fn open_read_with_buffer_size(stream: R, buffer_size: usize) -> io::Result<StreamingSpectrumIterator<C, D, MultiLayerSpectrum<C, D>, Self>> {
         let mut stream = PreBufferedStream::new_with_buffer_size(stream, buffer_size)?;
         let (fmt, gzipped) = infer_from_stream(&mut stream)?;
@@ -415,16 +492,7 @@ impl<R: io::Read,
             return Err(io::Error::new(io::ErrorKind::Unsupported, "This method does not support gzipped streams"))
         }
 
-        let reader = match fmt {
-            #[cfg(feature = "mgf")]
-            MassSpectrometryFormat::MGF => Self::MGF(MGFReaderType::new(stream)),
-            #[cfg(feature = "mzml")]
-            MassSpectrometryFormat::MzML => Self::MzML(MzMLReaderType::new(stream)),
-            _ => {
-                return Err(io::Error::new(io::ErrorKind::Unsupported, format!("This method does not support {fmt}")))
-            }
-        };
-        Ok(StreamingSpectrumIterator::new(reader))
+        MZReaderType::create_from_prebuffered_reader(stream, fmt)
     }
 }
 
@@ -766,7 +834,7 @@ mod async_impl {
     use crate::io::mgf::AsyncMGFReaderType;
     #[cfg(feature = "mzml")]
     use crate::io::mzml::AsyncMzMLReaderType;
-    use crate::io::traits::{AsyncMZFileReader, AsyncRandomAccessSpectrumIterator, AsyncSpectrumSource};
+    use crate::io::traits::{AsyncRandomAccessSpectrumIterator, AsyncSpectrumSource};
 
     #[cfg(feature = "thermo")]
     use crate::io::thermo::AsyncThermoRawReaderType;
@@ -955,7 +1023,7 @@ mod async_impl {
 
     #[cfg(feature = "async")]
     impl<C: CentroidLike + From<CentroidPeak> + BuildFromArrayMap + Send + Sync + 'static,
-         D: DeconvolutedCentroidLike + From<DeconvolutedPeak> + BuildFromArrayMap + Send + Sync + 'static> AsyncMZFileReader<C, D, MultiLayerSpectrum<C,D>> for AsyncMZReaderType<tokio::fs::File, C, D> {
+         D: DeconvolutedCentroidLike + From<DeconvolutedPeak> + BuildFromArrayMap + Send + Sync + 'static> crate::io::traits::AsyncMZFileReader<C, D, MultiLayerSpectrum<C,D>> for AsyncMZReaderType<tokio::fs::File, C, D> {
 
         async fn construct_index_from_stream(&mut self) -> u64 {
             amsfmt_dispatch!(self, reader, reader.construct_index_from_stream().await)
@@ -1057,6 +1125,8 @@ mod async_impl {
         #[cfg(feature = "async")]
         /// Create a reader from a file on the local file system denoted by `path`.
         pub async fn from_path<P: AsRef<Path>>(self, path: P) -> io::Result<AsyncMZReaderType<tokio::fs::File, C, D>> {
+            use crate::io::traits::AsyncMZFileReader;
+
             let mut reader = AsyncMZReaderType::open_path(path::PathBuf::from(path.as_ref())).await?;
             reader.set_detail_level(self.detail_level);
             Ok(reader)
@@ -1246,6 +1316,8 @@ impl<R: io::Read + io::Seek, C: FeatureLike<MZ, IonMobility>, D: FeatureLike<Mas
 
 #[cfg(test)]
 mod test {
+    #[allow(unused)]
+    use crate::prelude::*;
     use super::*;
 
     #[test]
@@ -1256,8 +1328,20 @@ mod test {
             assert_eq!(reader.as_format(), MassSpectrometryFormat::BrukerTDF);
             eprintln!("{}", reader.len());
             let s = reader.get_spectrum_by_index(0).unwrap();
-            assert!(s.peaks.is_some());
+            assert!(s.has_ion_mobility_dimension());
         }
+        Ok(())
+    }
+
+    #[test]
+    fn test_gzipped_read_seek() -> io::Result<()> {
+        let stream = fs::File::open("test/data/20200204_BU_8B8egg_1ug_uL_7charges_60_min_Slot2-11_1_244.mzML.gz")?;
+        let reader: MZReaderType<Box<dyn SeekRead>> = MZReader::<fs::File>::open_read_seek_generic(Box::new(stream) as Box<dyn SeekRead>)?;
+        assert_eq!(reader.len(), 103);
+
+        let stream = fs::File::open("test/data/20200204_BU_8B8egg_1ug_uL_7charges_60_min_Slot2-11_1_244.mzML.gz")?;
+        let reader = MZReader::<fs::File>::open_gzipped_read_seek(stream)?;
+        assert_eq!(reader.len(), 103);
         Ok(())
     }
 }