+++ src/io/mzml/mod.rs
@@ -19,7 +19,7 @@ mod reader;
 mod reading_shared;
 mod writer;
 
-#[cfg(feature = "async")]
+#[cfg(feature = "async_partial")]
 mod async_reader;
 
 pub use reading_shared::{
@@ -39,7 +39,7 @@ pub(crate) use crate::io::mzml::reader::is_mzml;
 
 pub use crate::io::mzml::writer::{MzMLWriter, MzMLWriterState, MzMLWriterType, MzMLWriterError};
 
-#[cfg(feature = "async")]
+#[cfg(feature = "async_partial")]
 pub use crate::io::mzml::async_reader::{
     MzMLReader as AsyncMzMLReader, MzMLReaderType as AsyncMzMLReaderType,
 };